refactor: Move types to a single index.d.ts file
