--- schemas/ts/Spec.ts
@@ -1,5 +1,0 @@
-import { Command } from './Command';
-
-export interface Spec extends Command {
-
-}
