--- schemas/ts/Suggestion.ts
@@ -1,6 +1,0 @@
-import { NamedCompletion } from './fig';
-
-export interface Suggestion extends NamedCompletion {
-
-    type?: string;
-}
